+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Notable changes are documented in this file, whose format follows [Keep a Change
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix native macOS/tvOS/watchOS test task dependency conflict with Compose Multiplatform resource assembly. ([#279](https://github.com/goncalossilva/kotlinx-resources/issues/279))
+
 ## [0.15.0] - 2026-02-09
 
 ### Added